Respect `disable_mcp_select` when playing discs
Due to the MemCard Pro code running slightly differently, when playing a disc with enhanced passthrough, the MemCard Pro would always be provided with the game’s title ID (and switch to a bespoke card), irrespective of the `disable_mcp_select` setting.

This adds an additional check to ensure that that setting is definitely adhered to.